backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.3",
    "greenlet>=3.4.0",
    "httpx>=0.28.1",
    "lxml>=6.0.3",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "tldextract>=5.3.1",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.2",
]
